+++ test/ttmlir/Dialect/TTNN/quantization/simple_requantize.mlir
@@ -1,9 +1,9 @@
 // RUN: ttmlir-opt --ttir-to-ttnn-backend-pipeline %s | FileCheck %s
 module {
-  func.func @forward(%arg0: tensor<1x3x320x320x!quant.uniform<i8:f32, 0.1>>) -> tensor<1x3x320x320x!quant.uniform<i8:f32, 0.2>> {
-    %0 = tensor.empty() : tensor<1x3x320x320x!quant.uniform<i8:f32, 0.2>>
-    %1 = "ttir.requantize"(%arg0, %0) : (tensor<1x3x320x320x!quant.uniform<i8:f32, 0.1>>, tensor<1x3x320x320x!quant.uniform<i8:f32, 0.2>>) -> tensor<1x3x320x320x!quant.uniform<i8:f32, 0.2>>
+  func.func @forward(%arg0: tensor<1x3x320x320x!quant.uniform<i32:f32, 0.1>>) -> tensor<1x3x320x320x!quant.uniform<i32:f32, 0.2>> {
+    %0 = tensor.empty() : tensor<1x3x320x320x!quant.uniform<i32:f32, 0.2>>
+    %1 = "ttir.requantize"(%arg0, %0) : (tensor<1x3x320x320x!quant.uniform<i32:f32, 0.1>>, tensor<1x3x320x320x!quant.uniform<i32:f32, 0.2>>) -> tensor<1x3x320x320x!quant.uniform<i32:f32, 0.2>>
     // CHECK: "ttnn.requantize"
-    return %1 : tensor<1x3x320x320x!quant.uniform<i8:f32, 0.2>>
+    return %1 : tensor<1x3x320x320x!quant.uniform<i32:f32, 0.2>>
   }
 }
