+++ test/ttmlir/Silicon/TTNN/n150/quantization/simple_dequantize.mlir
@@ -3,9 +3,9 @@
 // RUN: ttmlir-translate --ttnn-to-flatbuffer %t.mlir > %t.ttnn
 
 module {
-  func.func @forward(%arg0: tensor<1x3x320x320x!quant.uniform<i8:f32, 0.1>>) -> tensor<1x3x320x320xf32> {
+  func.func @forward(%arg0: tensor<1x3x320x320x!quant.uniform<i32:f32, 0.1>>) -> tensor<1x3x320x320xf32> {
     %0 = tensor.empty() : tensor<1x3x320x320xf32>
-    %1 = "ttir.dequantize"(%arg0, %0) : (tensor<1x3x320x320x!quant.uniform<i8:f32, 0.1>>, tensor<1x3x320x320xf32>) -> tensor<1x3x320x320xf32>
+    %1 = "ttir.dequantize"(%arg0, %0) : (tensor<1x3x320x320x!quant.uniform<i32:f32, 0.1>>, tensor<1x3x320x320xf32>) -> tensor<1x3x320x320xf32>
     // CHECK: "ttnn.dequantize"
     return %1 : tensor<1x3x320x320xf32>
   }
